Now a Card owns a FileSystem instance.
Before it was a reference, but this is not always easy to deal with so
I decided to avoid having a lifetime and my conclusion is it's better to
simply make Card owner of his own FileSystem instance.
Also it does not make sense to duplicate a Card instance so Clone and Copy will never be required for FileSystem.